Disable IPv6 for endpoints in '--ipv6=false' networks.
No IPAM IPv6 address is given to an interface in a network with
'--ipv6=false', but the kernel would assign a link-local address and,
in a macvlan/ipvlan network, the interface may get a SLAAC-assigned
address.

So, disable IPv6 on the interface to avoid that.

Signed-off-by: Rob Murray <rob.murray@docker.com>
(cherry picked from commit 841c4c8057bcf5317d6565875595a3f0c046e3fa)
Signed-off-by: Austin Vazquez <macedonv@amazon.com>